refactor: move client-side websocket
