+++ src/client/store.js
@@ -1,26 +1,10 @@
-import WS from 'ws';
 import { applyMiddleware, createStore, combineReducers } from 'redux';
 
 import messages from './reducers/chat';
 import route from './reducers/navigation';
 import game from '../common/game';
 
-const client = new WS('ws://localhost:3000');
-
-function send(type, payload, meta) {
-  const message = JSON.stringify({
-    type,
-    payload,
-    meta,
-  });
-
-  client.send(message);
-}
-
-client.onmessage = event => {
-  const action = JSON.parse(event.data);
-  store.dispatch(action);
-};
+import websocket from './middlewares/websocket';
 
 const middlewares = [
   store => next => action => {
@@ -34,12 +18,7 @@ const middlewares = [
     }
     next(action);
   },
-  store => next => action => {
-    if (action.meta && action.meta.pending) {
-      return send(action.type, action.payload, action.meta);
-    }
-    return next(action);
-  }
+  websocket,
 ];
 
 const store = applyMiddleware(...middlewares)(createStore)(combineReducers({
